+++ source/extensions/omni.isaac.ros2_bridge/backend/humble/Ros2HumbleService.cpp
@@ -11,7 +11,10 @@
 #include <include/Ros2Macros.h>
 #include <rcl/rcl.h>
 
-Ros2ServiceHumble::Ros2ServiceHumble(Ros2NodeBase* node, const char* service_name, const void* type)
+Ros2ServiceHumble::Ros2ServiceHumble(Ros2NodeBase* node,
+                                     const char* service_name,
+                                     const void* type,
+                                     const Ros2QoSProfile& qos)
     : mNode(node), wait_set_initialized(false)
 {
     mService = std::shared_ptr<rcl_service_t>(new rcl_service_t,
@@ -29,6 +32,7 @@ Ros2ServiceHumble::Ros2ServiceHumble(Ros2NodeBase* node, const char* service_nam
                                               });
     (*mService) = rcl_get_zero_initialized_service();
     rcl_service_options_t srv_ops = rcl_service_get_default_options();
+    srv_ops.qos = Ros2QoSProfileHumbleConverter::convert(qos);
     // srv_ops.qos.depth = history_depth;
 
     // rcl_service_default_options srv_ops = {