+++ source/extensions/omni.isaac.ros2_bridge/include/Ros2Factory.h
@@ -479,7 +479,10 @@ public:
                                                              const char* topic_name,
                                                              const void* type,
                                                              const Ros2QoSProfile& qos) = 0;
-    virtual std::shared_ptr<Ros2Service> CreateService(Ros2NodeBase* node, const char* service_name, const void* type) = 0;
+    virtual std::shared_ptr<Ros2Service> CreateService(Ros2NodeBase* node,
+                                                       const char* service_name,
+                                                       const void* type,
+                                                       const Ros2QoSProfile& qos) = 0;
 
     virtual std::shared_ptr<Ros2ClockMessage> CreateClockMessage() = 0;