+++ source/extensions/omni.isaac.ros2_bridge/nodes/OgnROS2PublishBbox3D.cpp
@@ -67,7 +67,19 @@ public:
             }
 
             Ros2QoSProfile qos;
-            qos.depth = db.inputs.queueSize();
+
+            const std::string& qosProfile = db.inputs.qosProfile();
+            if (qosProfile == "")
+            {
+                qos.depth = db.inputs.queueSize();
+            }
+            else
+            {
+                if (!jsonToRos2QoSProfile(qos, qosProfile))
+                {
+                    return false;
+                }
+            }
             state.mPublisher = state.mFactory->CreatePublisher(
                 state.mNodeHandle.get(), fullTopicName.c_str(), state.mMessage->getTypeSupportHandle(), qos);
             state.mFrameId = db.inputs.frameId();